# swagger.json
{"openapi":"3.0.0","paths":{"/streaming":{"get":{"operationId":"StreamingController_streaming","parameters":[],"responses":{"200":{"description":""}}}},"/auth/confirm-authentication":{"get":{"operationId":"AuthController_confirmAuthentication","summary":"激活账号","parameters":[{"name":"token","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["认证控制器"]}},"/auth/login":{"post":{"operationId":"AuthController_login","summary":"登录","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["认证控制器"]}},"/auth/github/{token}":{"get":{"operationId":"AuthController_github","summary":"Github登录","parameters":[],"responses":{"200":{"description":""}},"tags":["认证控制器"]}},"/auth/alipay/{token}":{"get":{"operationId":"AuthController_alipay","summary":"支付宝登录","parameters":[],"responses":{"200":{"description":""}},"tags":["认证控制器"]}},"/auth/logout":{"get":{"operationId":"AuthController_logout","summary":"登出","parameters":[],"responses":{"200":{"description":""}},"tags":["认证控制器"]}},"/account/register":{"post":{"operationId":"AccountController_register","summary":"注册账号","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserDto"}}}},"responses":{"201":{"description":""}},"tags":["账户控制器"]}},"/account/reset-password":{"post":{"operationId":"AccountController_resetPassword","summary":"重置密码","parameters":[],"responses":{"201":{"description":""}},"tags":["账户控制器"]}}},"info":{"title":"Webcaster","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{"accountIdentifier":{"type":"string","description":"用户名/邮箱/手机号","example":"username@example.com"},"password":{"type":"string","description":"密码"}},"required":["accountIdentifier","password"]},"RegisterUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]}}}}